+++ lib/bsev_widget.dart
@@ -1,25 +1,24 @@
 import 'package:bsev/bloc_base.dart';
-import 'package:bsev/bloc_view.dart';
 import 'package:bsev/communication_base.dart';
 import 'package:bsev/events_base.dart';
+import 'package:bsev/util.dart';
 import 'package:flutter/widgets.dart';
-
-import 'injector.dart';
 
 typedef AsyncWidgetBuilder<S> = Widget Function(
     BuildContext context, S communication);
 
-typedef ReceiveEventCallBack<S> = Function(EventsBase event, S communication);
+typedef ReceiveEventCallBack = Function(
+    EventsBase event, CommunicationBase communication);
 
 // ignore: must_be_immutable
 class Bsev<B extends BlocBase, S extends CommunicationBase>
     extends StatefulWidget {
   final dynamic dataToBloc;
   final AsyncWidgetBuilder<S> builder;
-  final ReceiveEventCallBack<S> eventReceiver;
+  final ReceiveEventCallBack eventReceiver;
 
   AsyncWidgetBuilder<CommunicationBase> _builderInner;
-  ReceiveEventCallBack<CommunicationBase> _eventReceiverInner;
+  ReceiveEventCallBack _eventReceiverInner;
 
   Bsev({Key key, @required this.builder, this.eventReceiver, this.dataToBloc})
       : super(key: key) {
@@ -31,43 +30,34 @@ class Bsev<B extends BlocBase, S extends CommunicationBase>
 
   void _confBuilders() {
     _builderInner = (context, communication) => builder(context, communication);
-    if (eventReceiver != null) {
-      _eventReceiverInner = (event, communication) {
-        return eventReceiver(event, communication);
-      };
-    }
+    _eventReceiverInner = (event, communication) {
+      if (eventReceiver != null) eventReceiver(event, communication);
+    };
   }
 }
 
 class _BsevState<B extends BlocBase, S extends CommunicationBase>
-    extends State<Bsev> implements BlocView<EventsBase> {
-  B _bloc;
-
-  @override
-  void eventReceiver(EventsBase event) {
-    if (widget._eventReceiverInner != null && mounted) {
-      widget._eventReceiverInner(event, _bloc.communication);
-    }
-  }
+    extends State<Bsev> {
+  S _communication;
 
   @override
   void initState() {
-    _bloc = getDependency<B>();
-    _bloc.data = widget.dataToBloc;
-    _bloc.setCommunication(getDependency<S>());
-    _bloc.communication.setView(this);
-    WidgetsBinding.instance.addPostFrameCallback((_) => _bloc.initView());
+    _communication = buildBsevCommunication<B, S>(
+      dataToBloc: widget.dataToBloc,
+      eventReceiver: widget._eventReceiverInner,
+    );
     super.initState();
   }
 
   @override
   void dispose() {
-    _bloc.dispose();
+    _communication.dispose();
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
-    return widget._builderInner(context, _bloc.communication);
+    _communication.initView();
+    return widget._builderInner(context, _communication);
   }
 }